Server fully created

+++ src/modules/users/user.service.ts
@@ -1,4 +1,5 @@
 import pool from "../../config/db";
+import { isAuthorized } from "../../middleware/auth";
 
 // get all users
 const getUser = async () => {
@@ -7,8 +8,16 @@ const getUser = async () => {
 }
 
 // update User
-const updateUser = async (payload: Record<string, unknown>, id: string) => {
+const updateUser = async (payload: Record<string, unknown>, id: string, user: any) => {
     const keys = Object.keys(payload);
+
+    // check is the user authentic
+    // console.log(id, user.userId, "isSame");
+    // if (user.role !== "admin" && id != user.userId) {
+    //     throw new Error("You are not allowed to update this profile");
+    // }
+    isAuthorized(user, id)
+
 
     if (keys.length === 0) {
         throw new Error("No fields provided to update");
@@ -36,6 +45,16 @@ const updateUser = async (payload: Record<string, unknown>, id: string) => {
 
 // delete user
 const deleteUser = async (id: string) => {
+    // check nay booking exist by this user!
+    const bookingCheck = await pool.query(`SELECT * FROM bookings WHERE customer_id=$1`, [id]);
+    const bookingStatus = bookingCheck.rows[0].status;
+
+    console.log(bookingStatus);
+    if (bookingStatus === "active") {
+        throw new Error("Failed to delete. This user have an active booking")
+    }
+
+
     const result = await pool.query(`DELETE FROM users WHERE id=$1`, [id]);
     return result
 }
